+++ examples/session_cookie_headers/session_with_cookies_headers.pas
@@ -6,7 +6,6 @@ var
   Session: THttpSession;
   Response: TResponse;
 begin
-  Session.Init;
   Session.SetBaseURL('https://httpbin.org');
   Session.SetHeader('X-Test-Header', 'RequestFP');
   Session.SetCookie('mycookie', 'cookievalue');